Now, don't need to convert it, as it is parsed to the correct integer type

+++ src/main.rs
@@ -7,7 +7,7 @@ fn multiplicative_persistence(n: u32) -> Vec<u32> {
         let mut result: u32 = 1;
 
         for i in num_string.chars() {
-            result *= (i.to_string().parse::<u32>().expect("Couldn't parse char to digit")) as u32;
+            result *= i.to_string().parse::<u32>().expect("Couldn't parse char to digit");
         }
 
         step_vec.push(result);
